refactor legend to it's own component
