+++ packages/component-library/src/Scatterplot/Scatterplot.js
@@ -10,6 +10,7 @@ import {
 } from 'victory';
 
 import ChartContainer from '../ChartContainer';
+import SimpleLegend from '../SimpleLegend';
 import { dollars, numeric } from '../utils/formatters';
 import CivicVictoryTheme from '../VictoryTheme/VictoryThemeIndex';
 
@@ -42,48 +43,6 @@ const chartEvents = [
     },
   },
 ];
-
-const SimpleLegend = ({ legendData }) => {
-  const legendStyle = css`
-    font-family: 'Roboto Condensed', 'Helvetica Neue', Helvetica, sans-serif;
-    font-size: 14px;
-    font-weight: bold;
-    text-align: center;
-    margin: 10px 0 0 0;
-  `;
-
-  if (legendData.length) {
-    return (
-      <legend className={legendStyle}>
-        {legendData.map((group, idx) => (
-          <span
-            key={group.name}
-            className={css`
-              margin-left: 10px;
-            `}
-          >
-            <svg viewBox="0 0 10 10" width="10px">
-              <circle
-                cx="5"
-                cy="5"
-                r="5"
-                fill={CivicVictoryTheme.civic.group.colorScale[idx]}
-              />
-            </svg>
-            <span
-              className={css`
-                margin-left: 5px;
-              `}
-            >
-              {group.name}
-            </span>
-          </span>
-        ))}
-      </legend>
-    );
-  }
-  return null;
-};
 
 const getDefaultDomain = (data, x, y) => {
   const xValues = data.map(value => value.x);
