+++ packages/component-library/src/SimpleLegend/index.js
@@ -1,0 +1,1 @@
+export default from './SimpleLegend';
